move modules out of module folder
